repr: extract into crate
Part of an effort to reduce compile times. Extracting repr into its own
crate facilitates breaking the dependency between sql and dataflow.
