+++ src/materialize/Cargo.toml
@@ -21,8 +21,8 @@ harness = false
 avro-rs = "0.6.5"
 backtrace = "0.3.34"
 bincode = "1.1.4"
-bytes = "0.4"
 byteorder = "1.3"
+bytes = "0.4"
 cast = "0.2"
 ccsr = { path = "../ccsr" }
 differential-dataflow = { git = "https://github.com/TimelyDataflow/differential-dataflow.git" }
@@ -31,6 +31,7 @@ futures = "0.1"
 getopts = "0.2"
 hyper = "0.12"
 itertools = "0.8.0"
+jemallocator = "0.3.0"
 lazy_static = "1.3.0"
 log = "0.4"
 metastore = { path = "../metastore" }
@@ -38,14 +39,14 @@ num-bigint = "0.2.2"
 ordered-float = { version = "1.0.2", features = ["serde"] }
 ore = { path = "../ore" }
 rand = "0.7"
+rdkafka = "0.21"
 regex = "1.2.1"
 regex-syntax = "0.6.8"
+repr = { path = "../repr" }
 reqwest = { version = "0.9.19", features = ["trust-dns"] }
-rdkafka = "0.21"
 serde = { version = "1.0", features = ["derive"] }
 serde_derive = "1.0"
 serde_json = "1.0.40"
-serde_regex = "0.4.0"
 sha2 = "0.8"
 simplelog = "0.6"
 smallvec = "0.6"
@@ -54,11 +55,10 @@ state_machine_future = { git = "https://github.com/benesch/state_machine_future.
 timely = { git = "https://github.com/TimelyDataflow/timely-dataflow.git", features = ["bincode"] }
 tokio = "0.1"
 tokio-threadpool = "0.1"
+unicase = "2.4.0"
 url = "1.7.2"
 url_serde = "0.2.0"
 uuid = { version = "0.7.4", features = ["serde", "v4"] }
-jemallocator = "0.3.0"
-unicase = "2.4.0"
 
 [dev-dependencies]
 pretty_assertions = "0.6.1"
