+++ src/materialize/dataflow/logging/differential.rs
@@ -6,8 +6,8 @@
 use super::{DifferentialLog, LogVariant};
 use crate::dataflow::arrangement::KeysOnlyHandle;
 use crate::dataflow::types::Timestamp;
-use crate::repr::Datum;
 use differential_dataflow::logging::DifferentialEvent;
+use repr::Datum;
 use std::time::Duration;
 use timely::communication::Allocate;
 use timely::dataflow::operators::capture::EventLink;
