+++ src/materialize/lib.rs
@@ -15,7 +15,6 @@ pub mod glue;
 pub mod interchange;
 pub mod pgwire;
 pub mod queue;
-pub mod repr;
 pub mod server;
 pub mod sql;
