+++ src/sqllogictest/Cargo.toml
@@ -12,22 +12,23 @@ path = "main.rs"
 path = "lib.rs"
 
 [dependencies]
-itertools = "0.8.0"
-sqlparser = { git = "ssh://git@github.com/MaterializeInc/sqlparser.git" }
-timely = { git = "https://github.com/TimelyDataflow/timely-dataflow.git", features = ["bincode"] }
-materialize = { path = "../materialize" }
-ore = { path = "../ore" }
-getopts = "0.2"
-walkdir = "2"
-regex = "1"
-serde_json = "1"
-lazy_static = "1"
+byteorder = "*"
 failure = "0.1.5"
 futures = "0.1"
-uuid = "*"
+getopts = "0.2"
+itertools = "0.8.0"
+lazy_static = "1"
+materialize = { path = "../materialize" }
 md5 = "*"
-rusqlite = { version = "0.20", features = ["bundled"] }
-postgres = "0.16.0-rc.2"
-byteorder = "*"
 ordered-float = "*"
+ore = { path = "../ore" }
+postgres = "0.16.0-rc.2"
+regex = "1"
+repr = { path = "../repr" }
+rusqlite = { version = "0.20", features = ["bundled"] }
+serde_json = "1"
+sqlparser = { git = "ssh://git@github.com/MaterializeInc/sqlparser.git" }
+timely = { git = "https://github.com/TimelyDataflow/timely-dataflow.git", features = ["bincode"] }
+uuid = "*"
+walkdir = "2"
 whoami = "0.5.3"
