+++ src/sqllogictest/lib.rs
@@ -24,10 +24,10 @@ use uuid::Uuid;
 use materialize::dataflow;
 use materialize::dataflow::{Dataflow, LocalSourceConnector, Source, SourceConnector};
 use materialize::glue::*;
-use materialize::repr::{ColumnType, Datum};
 use materialize::sql::store::RemoveMode;
 use materialize::sql::{Planner, Session};
 use ore::collections::CollectionExt;
+use repr::{ColumnType, Datum};
 use sqlparser::ast::{ObjectType, Statement};
 use sqlparser::dialect::AnsiDialect;
 use sqlparser::parser::{Parser as SqlParser, ParserError as SqlParserError};
